backend/pyproject.toml
[project]
name = "notebook-lm"
version = "0.1.0"
description = "Add your description here"
readme = "../README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "assemblyai>=0.44.3",
    "crewai>=0.193.2",
    "crewai-tools>=0.73.1",
    "fastembed>=0.7.3",
    "firecrawl-py>=4.3.6",
    "ipykernel>=6.30.1",
    "TTS>=0.22.0",
    "qdrant-client>=1.7.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.2.0",
    "soundfile>=0.12.1",
    "scipy>=1.11.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "yt-dlp>=2024.12.13",
    "zep-cloud>=3.4.3",
    "zep-crewai>=1.1.1",
    "requests>=2.31.0",
    # Authentication dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",  # For sync database operations (password reset)
    "pydantic>=2.5.0",
    "email-validator>=2.0.0",
    "apex-saas-framework>=0.1.0",
    "nest-asyncio>=1.5.8",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "certifi>=2023.11.17",
    "httpx>=0.25.0",
    "alembic>=1.13.0",
]
